requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
backports.tarfile==1.2.0
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
docutils==0.21.2
fastapi==0.115.4
fastapi-socketio==0.0.10
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
h11==0.14.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
imutils==0.5.4
intel-openmp==2021.4.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
nh3==0.2.18
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.16.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pkginfo==1.10.0
protobuf==5.28.3
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.3
scikit-image==0.24.0
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tbb==2021.13.1
tifffile==2024.9.20
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.5
wsproto==1.2.0
zipp==3.20.2
